backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & RAG
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.3

# Utils
requests==2.31.0
numpy==1.26.2
pandas==2.1.3